suporte ao novo event handler
