+++ eventos/ready.js
@@ -1,4 +1,9 @@
 // Emitido quando o bot fica pronto
-module.exports = async () => {
-    client.log("bot", `${client.user.tag} pronto!`);
+module.exports = {
+    nome: "ready",
+    once: true, // Se deve ser executado apenas uma vez
+
+    async executar() {
+        client.log("bot", `${client.user.tag} pronto!`);
+    }
 }
