+++ modulos/eventos.js
@@ -1,16 +1,22 @@
 const fs = require("fs");
 
 module.exports = () => {
-    let eventos = []
     for (const arquivo of fs.readdirSync(client.dir + `/eventos/`)) {
         if (!arquivo.endsWith(".js")) continue
-        client.log("verbose", `Carregando evento ${arquivo}...`)
-        const evento = require(client.dir + `/eventos/${arquivo}`);
-        let nomeEvento = arquivo.split(".")[0];
 
-        eventos.push(nomeEvento)
+        try {
+            client.log("verbose", `Carregando evento ${arquivo}...`)
+            const evento = require(client.dir + `/eventos/${arquivo}`);
 
-        client.on(nomeEvento, evento.bind(null));
-        client.log("verbose", `Evento carregado ${nomeEvento}`)
+            if (evento.once) {
+                client.once(evento.nome, (...args) => evento.executar(...args));
+            } else {
+                client.on(evento.nome, (...args) => evento.executar(...args));
+            }
+
+            client.log("verbose", `Evento carregado ${evento.nome}`)
+        } catch (err) {
+            client.log("critico", err.stack)
+        }
     }
 }
